scr/functions/api-extract/requirements.txt
requests>=2.31.0
google-cloud-pubsub>=2.18.0
google-cloud-storage>=2.10.0